# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
    title: Customer Service API
    description: Microservicio para la gestión de clientes bancarios
    version: 1.0.0
    contact:
        name: NTT Data Bootcamp
servers:
    - url: http://localhost:8080
      description: Servidor de desarrollo
...
paths:
...
    /personal:
        get:
            tags:
                - Personal Customers
            summary: Obtener todos los clientes personales
            description: Retorna una lista de todos los clientes personales
            operationId: retriveAllPersonalCustomers
            responses:
                '200':
                    description: Lista de clientes personales obtenida exitosamente
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/CustomerResponse'
                            example:
                                -   id: "123e4567-e89b-12d3-a456-426614174000"
                                    name: "Juan Pérez"
                                    customerType: "PERSONAL"
                                    documentNumber: "12345678"
                                    email: "juan.perez@email.com"
                                    active: true
                                    createdAt: "2024-01-15T10:00:00Z"
                                -   id: "456e7890-e89b-12d3-a456-426614174000"
                                    name: "María García"
                                    customerType: "PERSONAL"
                                    documentNumber: "87654321"
                                    email: "maria.garcia@email.com"
                                    active: true
                                    createdAt: "2024-01-15T11:00:00Z"
                '500':
                    description: Error interno del servidor
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'

        post:
            tags:
                - Personal Customers
            summary: Crear cliente personal
            description: Crea un nuevo cliente personal
            operationId: createPersonalCustomer
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreatePersonalRequest'
                        example:
                            firstName: "Juan"
                            lastName: "Pérez"
                            documentNumber: "12345678"
                            documentType: "DNI"
                            email: "juan.perez@email.com"
                            phone: "+51 999888777"
                            street: "Av. Principal 123"
                            city: "Lima"
                            country: "Perú"
                            zipCode: "15001"
                            birthDate: "1990-05-15"
            responses:
                '200':
                    description: Cliente personal creado exitosamente
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CustomerResponse'
                '400':
                    description: Datos de entrada inválidos
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                '500':
                    description: Error interno del servidor
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
    /personal/{id}:
        get:
            tags:
                - Personal Customers
            summary: Obtener cliente personal por ID
            description: Obtiene un cliente personal específico por su ID único
            operationId: retrievePersonalCustomer
            parameters:
                -   name: id
                    in: path
                    required: true
                    description: ID único del cliente personal
                    schema:
                        type: string
                        format: uuid
                    example: "123e4567-e89b-12d3-a456-426614174000"
            responses:
                '200':
                    description: Cliente personal obtenido exitosamente
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CustomerResponse'
                '404':
                    description: Cliente personal no encontrado
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                '500':
                    description: Error interno del servidor
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
    /personal/document-type/{documentType}/document-number/{documentNumber}:
        get:
            tags:
                - Personal Customers
            summary: Obtener cliente personal por documento
            description: Obtiene un cliente personal por tipo y número de documento
            operationId: retrievePersonalCustomerByDocument
            parameters:
                -   name: documentType
                    in: path
                    required: true
                    description: Tipo de documento
                    schema:
                        type: string
                    example: "DNI"
                -   name: documentNumber
                    in: path
                    required: true
                    description: Número de documento
                    schema:
                        type: string
                        minLength: 8
                        maxLength: 20
                    example: "12345678"
            responses:
                '200':
                    description: Cliente personal obtenido exitosamente
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CustomerResponse'
                '404':
                    description: Cliente personal no encontrado
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                '500':
                    description: Error interno del servidor
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
    /personal/update/{id}:
        put:
            tags:
                - Personal Customers
            summary: Actualizar cliente personal
            description: Actualiza la información de un cliente personal existente
            operationId: updatePersonalCustomer
            parameters:
                -   name: id
                    in: path
                    required: true
                    description: ID único del cliente personal
                    schema:
                        type: string
                        format: uuid
                    example: "123e4567-e89b-12d3-a456-426614174000"
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdatePersonalRequest'
                        example:
                            firstName: "Juan Carlos"
                            lastName: "Pérez García"
                            email: "juan.perez@nuevoemail.com"
                            phone: "+51 999888777"
                            street: "Av. Nueva Dirección 789"
                            city: "Arequipa"
                            country: "Perú"
                            zipCode: "04001"
                            birthDate: "1990-05-15"
            responses:
                '200':
                    description: Cliente personal actualizado exitosamente
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CustomerResponse'
                '400':
                    description: Datos de entrada inválidos
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                '404':
                    description: Cliente personal no encontrado
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                '500':
                    description: Error interno del servidor
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
    /personal/desactivate/{id}:
        delete:
            tags:
                - Personal Customers
            summary: Desactivar cliente personal
            description: Desactiva un cliente personal (soft delete)
            operationId: desactivatePersonalCustomer
            parameters:
                -   name: id
                    in: path
                    required: true
                    description: ID único del cliente personal
                    schema:
                        type: string
                        format: uuid
                    example: "123e4567-e89b-12d3-a456-426614174000"
            responses:
                '200':
                    description: Cliente personal desactivado exitosamente
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CustomerResponse'
                '404':
                    description: Cliente personal no encontrado
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                '500':
                    description: Error interno del servidor
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
...
components:
    schemas:
        CreatePersonalRequest:
            type: object
            required:
                - firstName
                - lastName
                - documentNumber
                - documentType
                - email
                - phone
                - street
                - city
                - country
                - zipCode
                - birthDate
            properties:
                firstName:
                    type: string
                    minLength: 1
                    maxLength: 100
                    description: Nombre del cliente personal
                    example: "Juan"
                lastName:
                    type: string
                    minLength: 1
                    maxLength: 100
                    description: Apellido del cliente personal
                    example: "Pérez"
                documentNumber:
                    type: string
                    minLength: 8
                    maxLength: 20
                    description: Número de documento
                    example: "12345678"
                documentType:
                    type: string
                    enum: [ DNI, PASSPORT ]
                    description: Tipo de documento
                    example: "DNI"
                email:
                    type: string
                    format: email
                    description: Correo electrónico del cliente
                    example: "juan.perez@email.com"
                phone:
                    type: string
                    minLength: 7
                    maxLength: 20
                    description: Número de teléfono
                    example: "+51 999888777"
                street:
                    type: string
                    minLength: 1
                    maxLength: 200
                    description: Dirección de la calle
                    example: "Av. Principal 123"
                city:
                    type: string
                    minLength: 1
                    maxLength: 100
                    description: Ciudad
                    example: "Lima"
                country:
                    type: string
                    minLength: 1
                    maxLength: 100
                    description: País
                    example: "Perú"
                zipCode:
                    type: string
                    minLength: 3
                    maxLength: 10
                    description: Código postal
                    example: "15001"
                birthDate:
                    type: string
                    format: date
                    description: Fecha de nacimiento
                    example: "1990-05-15"
        UpdatePersonalRequest:
            type: object
            properties:
                firstName:
                    type: string
                    minLength: 1
                    maxLength: 100
                    description: Nombre del cliente personal
                    example: "Juan Carlos"
                lastName:
                    type: string
                    minLength: 1
                    maxLength: 100
                    description: Apellido del cliente personal
                    example: "Pérez García"
                email:
                    type: string
                    format: email
                    description: Correo electrónico del cliente
                    example: "juan.perez@nuevoemail.com"
                phone:
                    type: string
                    minLength: 7
                    maxLength: 20
                    description: Número de teléfono
                    example: "+51 999888777"
                street:
                    type: string
                    minLength: 1
                    maxLength: 200
                    description: Dirección de la calle
                    example: "Av. Nueva Dirección 789"
                city:
                    type: string
                    minLength: 1
                    maxLength: 100
                    description: Ciudad
                    example: "Arequipa"
                country:
                    type: string
                    minLength: 1
                    maxLength: 100
                    description: País
                    example: "Perú"
                zipCode:
                    type: string
                    minLength: 3
                    maxLength: 10
                    description: Código postal
                    example: "04001"
                birthDate:
                    type: string
                    format: date
                    description: Fecha de nacimiento
                    example: "1990-05-15"
...
        CustomerResponse:
            type: object
            properties:
                id:
                    type: string
                    description: ID único del cliente
                    example: "987fcdeb-51a2-43d1-b789-123456789abc"
                name:
                    type: string
                    description: Nombre completo del cliente
                    example: "Empresa ABC SAC"
                customerType:
                    type: string
                    description: Tipo de cliente
                    example: "ENTERPRISE"
                documentNumber:
                    type: string
                    description: Número de documento
                    example: "20123456789"
                email:
                    type: string
                    format: email
                    description: Correo electrónico
                    example: "contacto@empresaabc.com"
                active:
                    type: boolean
                    description: Estado activo del cliente
                    example: true
                createdAt:
                    type: string
                    format: date-time
                    pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}$'
                    description: Fecha de creación
                    example: "2024-01-15T10:00:00Z"
        ErrorResponse:
            type: object
            properties:
                timestamp:
                    type: string
                    format: date-time
                    description: Timestamp del error
                    example: "2024-01-15T10:00:00Z"
                status:
                    type: integer
                    description: Código de estado HTTP
                    example: 400
                error:
                    type: string
                    description: Tipo de error
                    example: "Bad Request"
                message:
                    type: string
                    description: Mensaje descriptivo del error
                    example: "Business name is required for enterprise customer"
                path:
                    type: string
                    description: Ruta donde ocurrió el error
                    example: "/enterprise"